# docs/swagger/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "RBAC scaffolding based on Gin + Gorm + Casbin + Dig.",
        "title": "cappuccino",
        "contact": {},
        "license": {},
        "version": "5.2.1"
    },
    "host": "127.0.0.1:1207",
    "basePath": "/",
    "paths": {
...
        "/api/v1/pub/current/menutree": {
            "get": {
                "tags": [
                    "登录管理"
                ],
                "summary": "查询当前用户菜单树",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "查询结果：{list:菜单树}",
                        "schema": {
                            "$ref": "#/definitions/schema.Menu"
                        }
                    },
                    "401": {
                        "description": "{error:{code:0,message:未授权}}",
                        "schema": {
                            "$ref": "#/definitions/schema.HTTPError"
                        }
                    },
                    "500": {
                        "description": "{error:{code:0,message:服务器错误}}",
                        "schema": {
                            "$ref": "#/definitions/schema.HTTPError"
                        }
                    }
                }
            }
        },
        "/api/v1/pub/current/password": {
            "put": {
                "tags": [
                    "登录管理"
                ],
                "summary": "更新个人密码",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header"
                    },
                    {
                        "description": "请求参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schema.UpdatePasswordParam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{status:OK}",
                        "schema": {
                            "$ref": "#/definitions/schema.HTTPStatus"
                        }
                    },
                    "400": {
                        "description": "{error:{code:0,message:无效的请求参数}}",
                        "schema": {
                            "$ref": "#/definitions/schema.HTTPError"
                        }
                    },
                    "401": {
                        "description": "{error:{code:0,message:未授权}}",
                        "schema": {
                            "$ref": "#/definitions/schema.HTTPError"
                        }
                    },
                    "500": {
                        "description": "{error:{code:0,message:服务器错误}}",
                        "schema": {
                            "$ref": "#/definitions/schema.HTTPError"
                        }
                    }
                }
            }
        },
        "/api/v1/pub/current/user": {
            "get": {
                "tags": [
                    "登录管理"
                ],
                "summary": "获取当前用户信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schema.UserLoginInfo"
                        }
                    },
                    "401": {
                        "description": "{error:{code:0,message:未授权}}",
                        "schema": {
                            "$ref": "#/definitions/schema.HTTPError"
                        }
                    },
                    "500": {
                        "description": "{error:{code:0,message:服务器错误}}",
                        "schema": {
                            "$ref": "#/definitions/schema.HTTPError"
                        }
                    }
                }
            }
        },
        "/api/v1/pub/login": {
            "post": {
                "tags": [
                    "登录管理"
                ],
                "summary": "用户登录",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schema.LoginParam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schema.LoginTokenInfo"
                        }
                    },
                    "400": {
                        "description": "{error:{code:0,message:无效的请求参数}}",
                        "schema": {
                            "$ref": "#/definitions/schema.HTTPError"
                        }
                    },
                    "500": {
                        "description": "{error:{code:0,message:服务器错误}}",
                        "schema": {
                            "$ref": "#/definitions/schema.HTTPError"
                        }
                    }
                }
            }
        },
        "/api/v1/pub/login/captcha": {
            "get": {
                "produces": [
                    "image/png"
                ],
                "tags": [
                    "登录管理"
                ],
                "summary": "响应图形验证码",
                "parameters": [
                    {
                        "type": "string",
                        "description": "验证码ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "重新加载",
                        "name": "reload",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "图形验证码"
                    },
                    "400": {
                        "description": "{error:{code:0,message:无效的请求参数}}",
                        "schema": {
                            "$ref": "#/definitions/schema.HTTPError"
                        }
                    },
                    "500": {
                        "description": "{error:{code:0,message:服务器错误}}",
                        "schema": {
                            "$ref": "#/definitions/schema.HTTPError"
                        }
                    }
                }
            }
        },
...
        "/api/v1/pub/refresh-token": {
            "post": {
                "tags": [
                    "登录管理"
                ],
                "summary": "刷新令牌",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/schema.LoginTokenInfo"
                        }
                    },
                    "401": {
                        "description": "{error:{code:0,message:未授权}}",
                        "schema": {
                            "$ref": "#/definitions/schema.HTTPError"
                        }
                    },
                    "500": {
                        "description": "{error:{code:0,message:服务器错误}}",
                        "schema": {
                            "$ref": "#/definitions/schema.HTTPError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "schema.HTTPError": {
            "type": "object",
            "properties": {
                "error": {
                    "description": "错误项",
                    "type": "object",
                    "$ref": "#/definitions/schema.HTTPErrorItem"
                }
            }
        },
        "schema.HTTPErrorItem": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "错误码",
                    "type": "integer"
                },
                "message": {
                    "description": "错误信息",
                    "type": "string"
                }
            }
        },
        "schema.HTTPStatus": {
            "type": "object",
            "properties": {
                "status": {
                    "description": "状态(OK)",
                    "type": "string"
                }
            }
        },
...
        "schema.LoginParam": {
            "type": "object",
            "required": [
                "captcha_code",
                "captcha_id",
                "password",
                "user_name"
            ],
            "properties": {
                "captcha_code": {
                    "description": "验证码",
                    "type": "string"
                },
                "captcha_id": {
                    "description": "验证码ID",
                    "type": "string"
                },
                "password": {
                    "description": "密码(md5加密)",
                    "type": "string"
                },
                "user_name": {
                    "description": "用户名",
                    "type": "string"
                }
            }
        },
        "schema.LoginTokenInfo": {
            "type": "object",
            "properties": {
                "access_token": {
                    "description": "访问令牌",
                    "type": "string"
                },
                "expires_at": {
                    "description": "令牌到期时间戳",
                    "type": "integer"
                },
                "token_type": {
                    "description": "令牌类型",
                    "type": "string"
                }
            }
        },
        "schema.Menu": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "actions": {
                    "description": "动作列表",
                    "type": "object",
                    "$ref": "#/definitions/schema.MenuActions"
                },
                "created_at": {
                    "description": "创建时间",
                    "type": "string"
                },
                "creator": {
                    "description": "创建者",
                    "type": "string"
                },
                "hidden": {
                    "description": "隐藏菜单(0:不隐藏 1:隐藏)",
                    "type": "integer"
                },
                "icon": {
                    "description": "菜单图标",
                    "type": "string"
                },
                "name": {
                    "description": "菜单名称",
                    "type": "string"
                },
                "parent_id": {
                    "description": "父级ID",
                    "type": "string"
                },
                "parent_path": {
                    "description": "父级路径",
                    "type": "string"
                },
                "record_id": {
                    "description": "记录ID",
                    "type": "string"
                },
                "resources": {
                    "description": "资源列表",
                    "type": "object",
                    "$ref": "#/definitions/schema.MenuResources"
                },
                "router": {
                    "description": "访问路由",
                    "type": "string"
                },
                "sequence": {
                    "description": "排序值",
                    "type": "integer"
                }
            }
        },
        "schema.MenuAction": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "动作编号",
                    "type": "string"
                },
                "name": {
                    "description": "动作名称",
                    "type": "string"
                }
            }
        },
        "schema.MenuActions": {
            "type": "array",
            "items": {
                "$ref": "#/definitions/schema.MenuAction"
            }
        },
        "schema.MenuResource": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "资源编号",
                    "type": "string"
                },
                "method": {
                    "description": "请求方式",
                    "type": "string"
                },
                "name": {
                    "description": "资源名称",
                    "type": "string"
                },
                "path": {
                    "description": "请求路径",
                    "type": "string"
                }
            }
        },
        "schema.MenuResources": {
            "type": "array",
            "items": {
                "$ref": "#/definitions/schema.MenuResource"
            }
        },
...
        "schema.UpdatePasswordParam": {
            "type": "object",
            "required": [
                "new_password",
                "old_password"
            ],
            "properties": {
                "new_password": {
                    "description": "新密码(md5加密)",
                    "type": "string"
                },
                "old_password": {
                    "description": "旧密码(md5加密)",
                    "type": "string"
                }
            }
        },
...
        "schema.UserLoginInfo": {
            "type": "object",
            "properties": {
                "real_name": {
                    "description": "真实姓名",
                    "type": "string"
                },
                "role_names": {
                    "description": "角色名列表",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "user_name": {
                    "description": "用户名",
                    "type": "string"
                }
            }
        },
...
    }
}